Moved faceDownCards variable and function from TableauPile to KlondikeTableau as so far only Klondike games would need them. Changed autoComplete and recordHistory to abstract as they would need slightly different implementations depending on game.
