+++ app/src/main/java/com/heyzeusv/solitaire/ui/AustralianPatienceViewModel.kt
@@ -1,8 +1,13 @@
 package com.heyzeusv.solitaire.ui
 
+import androidx.compose.runtime.snapshots.Snapshot
 import com.heyzeusv.solitaire.data.AustralianPatienceTableau
+import com.heyzeusv.solitaire.data.Foundation
+import com.heyzeusv.solitaire.data.PileHistory
 import com.heyzeusv.solitaire.data.ShuffleSeed
+import com.heyzeusv.solitaire.data.Stock
 import com.heyzeusv.solitaire.data.TableauPile
+import com.heyzeusv.solitaire.data.Waste
 import com.heyzeusv.solitaire.util.ResetOptions
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
@@ -16,7 +21,7 @@ import javax.inject.Inject
 @HiltViewModel
 class AustralianPatienceViewModel @Inject constructor(
     ss: ShuffleSeed
-) : GameViewModel() {
+) : GameViewModel(ss) {
 
     override val _tableau: MutableList<TableauPile> = MutableList(7) { AustralianPatienceTableau() }
 
@@ -30,8 +35,28 @@ class AustralianPatienceViewModel @Inject constructor(
         }
     }
 
+    override fun autoComplete() {
+        TODO("Not yet implemented")
+    }
+
+    /**
+     *  Takes a [Snapshot] of current StateObject values and stores them in a [PileHistory] object. We
+     *  then immediately dispose of the [Snapshot] to avoid memory leaks.
+     */
+    override fun recordHistory() {
+        val currentSnapshot = Snapshot.takeMutableSnapshot()
+        currentSnapshot.enter {
+            currentStep = PileHistory(
+                stock = Stock(_stock.pile),
+                waste = Waste(_waste.pile),
+                foundation = _foundation.map { Foundation(it.suit, it.pile) },
+                tableau = _tableau.map { AustralianPatienceTableau(it.pile) }
+            )
+        }
+        currentSnapshot.dispose()
+    }
+
     init {
-        mSS = ss
         resetAll(ResetOptions.NEW)
     }
 }
